pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "korely-graphrag"
version = "0.1.0"
description = "MCP-native second brain with an automatic entity graph over your markdown notes."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
    { name = "Massimiliano Martella" },
]
keywords = ["mcp", "rag", "knowledge-graph", "second-brain", "gemini", "pgvector", "obsidian"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastmcp>=0.2.0",
    "google-genai>=0.3.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "pgvector>=0.2.5",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "click>=8.1",
    "uvicorn>=0.30",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

[project.scripts]
korely-graphrag = "korely_graphrag.cli:main"

[project.urls]
Homepage = "https://github.com/verdana86/korely-graphrag"
Issues = "https://github.com/verdana86/korely-graphrag/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/korely_graphrag"]

[tool.ruff]
line-length = 100
target-version = "py311"
